+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   during generation with vLLM.
 - Avoids excessive disk usage by not caching processed datasets to disk, as we are
   never using the cached versions anyway.
+- In sequence classification tasks, we now only strip the prompts if the model's
+  tokenizer includes a prefix space when tokenizing the labels.
 
 ### Changed
 - Swapped primary/secondary metrics for the multiple choice tasks, where we now set MCC